Highlight "Previous" and "Next" Lesson Nav buttons
Move lesson state to redux, with a "reselector"

The problem this solves: Lesson was effectively reading the state and
then mutating it as a result of its new values.

The solution: a "selector" turns the state into something useful in all
the modules that test it.

No more stateHasLessonHighlight(). Now it's
lessonSelector(state).testHighlight({ type: 'Foo', ... }) ... which fits
better in mapStateToProps().

Came up with a mount()-based testing strategy that seems to balance
verbosity with usefulness.