Nix Moment and Chartbuilder
Chartbuilder: we only needed a few methods, not Flux and React and
whatnot. So I:

1. Untangled the Flux stuff and made it synchronous
2. Cleaned up our requirements tree
3. Copied the Chartbuilder source code that we need into
   assets/js/wfparameters/charts/chartbuilder
4. Removed dependencies (chartbuilder, chartbuilder-ui, react-tangle,
   and all the other things we evidently only added to make those work)
5. Tested

I discovered that Chartbuilder depends on sugar-date, which modifies the
global Date prototype. That's unacceptable. I rewrote the methods it
needed.

Next up: timediff. It's obsolete and huge (it depends on moment.js,
which is badly packaged so includes all locale data). Removing it was
easy: we never needed it in the first place.

Yay! Dev build is ~60% its former size.

Next we can upgrade React and use all its wonderful new features, like
portals and refs.