Integration-test almost all of Lesson 1
I ran into an enormous problem with drag-and-drop. I upgraded and
fiddled with browsers and spent 8h on the whole thing.

In the end, there are a few empty <div>s gone from our layout. Win!

Rather than waste any more time on this, I'll do #157439948
and #157439963 so we don't need drag-and-drop.