Show module "busy" status
[finishes #159117149] but not as described in the story because I don't
understand the story or the problem it tries to solve.

This solves the problem of "I don't see that Twitter is fetching." It
does it by putting a spinner on the Fetch button whenever the module is
busy (as determined by the server).

I think this solves our biggest outstanding responsiveness problem. Now
... what's left? Let's discuss (or accept this story).