Track lesson sections, not steps
[refs #160041803]

Intercom has an event-name limit, so we no longer create an event name
for every step of every lesson. Instead, we go one per section.

I've also added `help` in the event metadata, so it will be easy to
understand that the events are fired _before_ the step is finished, not
afterwards.