Execute full workflow with every edit
The good: when the user changes something, the whole workflow updates.

The bad: if the user changes two things, and each request goes to a
different server, then both servers may keep some of the workflow in
RAM as they race each other to the finish. (They won't overwrite one
another's results or duplicate work.) The worst case is:

1. Process A executes module 1, caching result and keeping it in RAM
2. Process B executes module 2, caching result and keeping it in RAM
3. Process A forgets module 1's output and loads module 2's output
... and they continue in lockstep.

Furthermore, if 10 users change params at once, we may keep 10 cached
interim results in RAM at once. (Previously we were bounded by the
number of web-serving threads; now we use the event loop.)

The solution is to execute in a background process. Today's commit
brings us much closer to that long-term solution.

The ugly: our scheduling is a bit ... overkill. With a proper
background process we could simplify tremendously. The primitives are:

* Re-execute Workflow (perhaps at delta X, starting module Y)
* Wait for WfModule result (perhaps at delta X) (and grab it before
  it gets overwritten)