plugin html: nix window.workbench.embeddata.
It's gotta go, because it encourages a race: if the module isn't
rendered when we load, its HTML has no way to know when the render
completes. (Hashchange may fire before its listener is installed.)

Module HTML should now use this pattern:

    function startLoading() { ... }

    window.addEventListener('hashchange', startLoading)
    startLoading()

With this change, the iframe will load new data on hashchange, for
_every_ hashchange. If it initially loads invalid data, it can correct
itself soon after.