Include fb-re2
This is so the wordcloud module can use Twittok's regex. Later we'll
probably want to rewrite some regexes using re2. In particular,
`formula` and `filter` ought to use re2 because that's what Google
Sheets uses so we'll be able to benefit from all that documentation.
(Not to mention: re2 is a zillion times faster.)