worker: die on DatabaseError
If a thread throws DatabaseError, its django.db.connections are in an
undefined state. We don't have a good place to fix that. (The
connections _ought_ to be guaranteed-clean by @database_sync_to_async,
but today on prod that didn't happen.)

Dying is a part of life! Workers will restart and reconnect; renders and
fetches won't be lost.