minio: connect on use, not when code loads
The previous behavior served two purposes: 1) crash immediately on prod when
minio is down, so we get notified of the error; and 2) auto-create buckets in
unit tests. But 1) is questionable anyway; and 2) is flawed. We auto-created
the bucket but never _cleared_ it between tests. Better is to create the
bucket in DbTestCase.setUp(). It's not terribly slow, and it prevents infinite
disk growth in dev mode.