Nix module_version in client
Now we always _pass_ a ModuleVersion to the client; but we _call_ it a
"Module". We nix module IDs entirely, in favor of slugs ("id_names").

This simplifies the Redux state and paves the way for an import/export
format. (Basically, import/export only work if the Redux state has no
IDs.)

There are lots of modifications in this commit. They are _all_
simplifications.

[refs #162704742]