wf_module.module_id_name: will replace .module_version_id
This way, the schema matches our intent: when you install a new ModuleVersion
every WfModule automatically switches to it.

This is _not_ a foreign key: it is possible to delete a Module without destroying
the user's data. Indeed, this is something we do on production: we delete an
in-use module if it isn't up to scratch.