importmodulefromgithub: do not derive link/author
The derived values get overridden when we re-import modules -- such as
when I migrated from `server_module` to `server_moduleversion`. This is
the sort of error that's unavoiable when external and derived values are
mashed together into a single data structure. Enough!

Since the server is missing all the old urls, I also nixed the 'imported
from another repo' error message. That message wasn't helping us in
practice.