minio+parquet: download all at once when possible
We're making too many S3 requests on production. The medium-term
solution is to store all header information in the database. Short-term,
we can avoid downloading a Parquet file piecemeal when we know we're
going to read the entire file.

This commit was a good idea, but it doesn't affect CachedRenderResult. I
think I'll go rethink CachedRenderResult now.