WIP refactoring render params.
This code _should_ load tab data when given a 'Tab' param. But it hasn't
been tested -- either in code or in unit tests. I'm committing it
anyway: this section of the codebase is all-new and sorta prototypical,
and I'm excited that it actually functions as a drop-in replacement --
even _without_ any actual tab params anywhere.