renderprep: prep multichartseries correctly
Previously, we were giving modules empty colnames, which led to
user-facing errors. Users had workarounds, but still -- it wasn't nice.

The error in process: we still have
ParamDType.omit_missing_table_columns(). That method is misleading
(because it seems like it's actually _used_ in render prep). I added a
comment; hopefully we'll soon be able to nix those methods and avoid
this sort of bug.