ValueSelect: calculate size automatically
[finishes #162109048]

It turns out react-virtualized-auto-sizer is not what we want. Instead,
we can roll our own component to calculate the height and maxHeight from
CSS. This makes styling easy, and it prevents the glitches on production
in which the height of the list item isn't the height of its contents
(so you can't, e.g., click anywhere on a <label> to select it).

Also, some naming conventions are getting clearer here: 'item' is a
value from the input data set, and 'value' is the user-supplied Array of
Strings that ValueSelect modifies. I didn't rename 100% of stuff, but I
renamed a bit.

Finally, we now report "Loading" and "Column contains no values" instead
of just leaving things empty.