react-select__menu: let Popper place it 'top' instead of 'bottom'
Previously, .react-select-menu-portal had (incorrect) height=0px because
.react-select__menu was position:absolute. So Popper couldn't do
what it was meant to do.

See also [refs #163066332] -- the expected behavior is for the module
stack to scroll down and for the menu to appear _below_ the prompt. This
commit doesn't interfere with that desired behavior. It only handles
the corner case that occurs when there isn't enough room below the
prompt.