converttotext: internal now; respect type formats
We don't have a solid way to pass code to modules yet. So we import
NumberFormatter in converttotext, which means we need to be internal.
(Normally, "things are under-designed" is a bad reason to internalize
a module. But converttotext _screams_ to be internal -- it's a core
feature -- so now is a great time to do it.)

There's a migration to change the module id_name: the original id_name
is incompatible with internal modules. Grr. So this deploy will be a
bit scary.