Migrate from minio-py to boto3
Why? Because minio's uploader would pre-read the entire file during
upload, leading to OutOfMemory. That's a rookie mistake, and it's always
been there. I explored GitHub for OutOfMemory-related issues and found a
lot; so I've judged that we should abandon this project because it seems
unequipped to detect or fix this problem -- which is, in my view, the
_entire point_ of writing an S3 client in the first place.

boto3 is flaws. It costs 36MB, which is nuts since we want nothing but
S3. Its exceptions seem inconsistent -- they're totally different for
head_object() than for get_object(). Oh, and there's an HTTP spec
violation that breaks minio: https://github.com/boto/boto3/issues/1341
so we have to monkey-patch that.

This should resolve OutOfMemory on production. That's a big deal.