Create separate DB-connections threadpool
This involves a new @database_sync_to_async decorator.

The main reason: we were leaking database connections in unit tests,
_non-deterministically_. This change makes the leaks deterministic.

Then I found and fixed the leaks. (They were all in the test suite, not
in the code.)

A side-effect: we should massively increase scalability, since we'll
hold fewer database connections. (The limit should be 2.)