wfmodule_moduledata endpoint: queue render if needed
Previously, the user had to actually load the Workflow page to kick off
a render of a module that has no cached data. Now, requesting the API
endpoint will start a render; so if the user retries a bunch of times on
empty result, sooner or later the result will be non-empty.