Refactor JS+CSS; add soon-to-be-deleted "Report" component
We have decided that the "Report" design is better implemented using
an iframe -- and the iframe needn't load React. The big win here is
that we now have a "Selected Pane" concept in WorkflowEditor -- instead
of just "Selected Tab". This lets us select a "Report" pseudo-tab.

TODO migrate the "Report" to plain HTML; make the "Report" React
component render the iframe.