WorkflowMetadata: pull out of WorkflowNavBar
We had two very different behaviors (one with a reducer, one without);
and they're two very different styles, as well (one in a list item, one
at the start of a page).

I haven't changed the CSS, so _that's_ still a mess. But at least the
<ShareButton> is a <ShareButton> -- and sharing from the "public" text
actually works again. (it regressed after #165837853)