parquet: read empty categoricals correctly
One of the following problems is on production, and I haven't figured
out which it is:

1. Maybe Parquet format doesn't store dtype of dictionary columns?
2. Maybe fastparquet doesn't write dtype of dictionary columns?
3. Maybe fastparquet doesn't read dtype of dictionary columns?

If the problem is that fastparquet writes incorrectly, that's sad; this
commit won't fix it, but since there's already a problem on production I
won't delve deeper today. Otherwise, the hack in this commit is correct:
when reading, convert all empty dictionaries to dtype=str.

Should fix error on production:

Exception in replace.render
Traceback (most recent call last):
  File "/app/server/models/loaded_module.py", line 151, in render
    out = self.render_impl(table, params, **kwargs)
  File "<Module replace.2723885>", line 82, in render
  File "<Module replace.2723885>", line 32, in process_table
  File "<Module replace.2723885>", line 37, in process_series
  File "<Module replace.2723885>", line 49, in _process_categorical
  File "/usr/local/lib/python3.7/site-packages/pandas/core/accessor.py", line 171, in __get__
    accessor_obj = self._accessor(obj)
  File "/usr/local/lib/python3.7/site-packages/pandas/core/strings.py", line 1796, in __init__
    self._validate(data)
  File "/usr/local/lib/python3.7/site-packages/pandas/core/strings.py", line 1834, in _validate
    raise AttributeError(message)
AttributeError: Can only use .str accessor with string values (i.e. inferred_type is 'string', 'unicode' or 'mixed')