views.upload: create ChangeParametersCommand on submit
[refs #166645919]

I'm testing out a change to Workflow.cooperative_lock(): its retval is
now a "WorkflowCooperativeLock" and it has an `after_commit(fn)` method
to queue the sending of Websockets deltas after commit. This makes views
easier to write in the common case: they can be decorated by a
cooperative lock.