server: use pyarrow instead of pandas
This is the start of a larger shift: Workbench components should
communicate with Arrow files (mmapped) rather than Parquet files (RAM).
`server` is a nice place to start because it uses pandas so very little
(aside from server.modules, which doesn't really belong in server).

This is more of a proof-of-concept than a useful change. It would be
nice to remove `server`'s dependency on Pandas, but currently pyarrow
depends on Pandas so there's not much point in nixing the dep.