Test message ID instead of text in SelectedRowsActions
The sample code that lingui provides for jest testing (and we use in test-utils) returns the IDs of messages instead of their translation.
This may be because otherwise each snapshot would need to contain the entire testing language catalogs and translation files would need to be up to date with the code before running (the corresponding) tests.