Never read using fastparquet. Avoids a segfault.
I don't quite understand the segfault, but it's something to do with:

1. fastparquet to read column as Categorical series (works)
2. encode as pyarrow (works)
3. call pyarrow.Table.from_arrays(...) (segfault!)

When we get pyarrow 0.15 we won't even be _tempted_ to do this
optimization. In the meantime ... the only reason I really optimized it
was because frontend was giving 502 Bad Gateway (timeout) on production
for general.csv ... and I think the `only_rows` param I added may be
enough to fix that problem.