rendercache.read_cached_render_result_pydict()
Now, when views.WfModule reads data for the user to view, we read
straight Parquet => Arrow => Python objects, one column at a time.
At any given time, in RAM, we have at most 1 column's worth of Arrow
data and the return value (200 Python objects per column). This is
an improvement over previously, when we converted the whole thing to
Arrow and then read it. ("Convert-then-read" might be cool if we had
a cache ... but we don't.)

The key win here is `only_rows`. I originally tried to make
`only_rows` a parameter to `convert_parquet_file_to_arrow()` ... but
it was difficult ... and ultimately unneeded.

We still have a high-RAM-usage function: the one that exports the
entire table as a CSV or JSON file. This was already a problem, and
it continues to be a problem. We'll get out-of-memory when users access
it.