Cache migrate_params() results.
On production, migrate_params() on the frontend was costing ~500ms per
module. That's astonishing (it's ~25ms on my dev machine). This quick
fix should bring our performance back where it belongs. (The problems
we've seen may return when loading a large, old workflow -- but to the
user it may seem like we're working hard because we're doing lots of
reading, which I guess we are.)

Incidentally: now, when serializing invalid params, we log and coerce.
Previously, we gave a 500 error. We've seen that 500 error often, and
never have I thought, "it would be worse if we didn't coerce." The user
sees a problem, yes; and we can end up losing users' data ... but at
least they can use Workbench.

It turns out we _write_ wf_module.params in very few places; so it's
easy to make all those actions invalidate the cache.