forkserver: bespoke, replacing Python's
This forkserver is thread-safe and asyncio-safe. It doesn't return
spurious "exit code 255" errors like we've seen on production. The logic
is much simpler.

Also, module processes are direct children of the calling process. That
makes it easier to reason about security and races; and it makes the
calling code simpler.

Unfortunately, `frontend` is failing. Something is reaping the child
processes for us, so we can't read the exit codes. I suspect Twisted
(which we ought not to even be using). So I'm committing what we have,
and now I'll embark on an investigation.