Use PR_SET_PROCESS_SUBREAPER correctly
I had totally misunderstood the calling convention. The *parent* process
needs to be marked subreaper. That, plus the existing lazy-loading of
staticmodules.registry, inspired me to create a `cjwstate.modules`
package. That's where our kernel singleton, static registry, and cached
loaded modules go. All components need to initialize the module system
during startup.

This is yet another step away from Django models -- we're taking the
things from that directory that don't use the database and we're putting
them in an un-Django place.

In the process, I experimented with Uvicorn. I like it. We should use
it.