forkserver.main: avoid "bad file descriptor" while exiting
We had a context manager for the socket ... so when the module process
raised an exception, it would try to close the socket while crashing.
But in the module, the socket's file descriptor is already closed.

The solution: no context managers. No try/catch. No finalizers.

os.fork() is confusing! :)