Add --clean option to lingui extract in bin/dev
When lingui scans the code, it comments out any messages of the catalog
that are not in the js code. Hence, messages already extracted by babel
will be first commented out and then readded by babel.
The --clean option makes lingui delete the message
instead of commenting it out; hence, duplication of messages is prevented.

This will not be required in version 3 of lingui,
which will give the ability to define multiple message catalogs,
some of which can be managed externally, that will be merged during its compilation step.
In this case, we would keep two separate message catalogs for source messages
and compile them both for js (with lingui)
and for babel (just merge its own source catalog with the source catalog of lingui).