Use django helpers in order to find and/or change active locale
When we add translations to contexts with no request
(such as lazy translations for form field placeholders),
we will need a method to find the currently active locale
(and, hence, a pairing one to set it when we know it).
The one used by django may or may not suit us,
however we definitely need something to start with.