pyarrow 0.15.0
Big changes:

* `/usr/bin/parquet-to-arrow`: convert to Arrow in an external process
* pyarrow.Table's "columns" are now ChunkedArray, not Column
* pyarrow reads Dict columns as Dict directly

There's a bug somewhere in pyarrow: when doing dict->dict conversion in
renderer, when rendercache contains a pyarrow-0.14.1 Parquet file written
late September, there's a read error:

Failure reading column: IOError: Arrow error: Invalid: Resize cannot downsize

... I'm trying to figure out the bug in Arrow. Maybe we have to live with
it? Anyway, there's an elegant fix in renderer: when we try rendering but the
input is a CorruptCacheError, backtrack and render the previous step.

I'm assuming the bug only happens when reading Arrow-0.14.1 Parquet files in
Arrow 0.15.0. (Maybe it was a bug in Arrow 0.14.1....) If the bug is internal
to Arrow 0.15.0, then let's revert this commit pending an Arrow bugfix.