Singleton chroot built with overlayfs
WE MUST HACK KUBERNETES DEPLOYMENTS TO MAKE THIS WORK

The problem this solves is: our previous chroot environment was slow to
build and was missing key files. We needed to hard-code every single
import imaginable in `cjwkernel/kernel.py` -- otherwise modules would
crash because the underlying .py files didn't exist. (If we'd copied all
.py files into the chroot, it would have taken a minute per module
run.)

There are two chroots: cjwkernel.chroot.EDITABLE_CHROOT (a singleton)
and cjwkernel.chroot.READONLY_CHROOT (re-entrant). migrate_params() and
compile() use READONLY_CHROOT, so we can run many modules concurrently.

The conceptual downside: we're only allowed one render() or fetch() at
a time, since they require EDITABLE_CHROOT. No problem: we _already_ only
run one at a time, to give our customers performance guarantees (e.g.,
"pay $x/s for 1 CPU, 2GB RAM"). I added a threading.Lock to enforce this;
by design, this is entirely redundant. (It would simply crash if we make
a terrible mistake elsewhere.)

The pragmatic downside: permissions. Kubernetes and Docker have various
security layers that prevent us from using overlayfs in an unprivileged
container. On dev, we add privileges. On production we'll need a sidecar
container running in privileged mode.

And because of this problem with privileges, it's really hard to "reset"
the chroot. The obvious way to "reset" the chroot is to delete all the
edits. But Linux's overlayfs filesystem gives undefined behavior if we
edit underlying layers. If we had privileges, we could simply unmount,
wipe edits, and remount. Lacking that, we have funky logic that _reads_
from the upper layer and _deletes_ from the overlayfs filesystem. The
danger: can a module actually leave a file behind on the filesystem? I
don't think so -- I think the deletion code would crash if it failed to
delete stuff. But if it somehow did manage to leak files, could that be
the start of an elaborate exploit? I don't think so ... but maybe I
lack imagination.

This revisiting of chroot is a precursor to adding a singleton network
namespace. I expect many similar problems to arise, and I'm glad to
have tackled most architectural problems before encountering them.