kubernetes: upgrade nginx
This seemed to be necessary as part of the v1.14 upgrade, but
I haven't explored why or whether I'm even right.

The new namespace is "ingress-nginx-v2" because of some stupid
disaster stuff on production. It ought to be "ingress-nginx." Fix
this when we split staging from production. (That'll be a great
day, because we'll be able to test cluster changes without taking
down our site.)