integrationtests: setup-chroots.sh in frontend+renderer+fetcher
This is closer to what we do on Kubernetes, where setup-chroots.sh
gets run in an init container. The difference is, in integration-test
the containers keep CAP_SYS_ADMIN after setting up chroots.

It's bad that CAP_SYS_ADMIN happens in integration-test and not
production; but how bad is this, *really*? Put it in perspective:
*all* our sandboxing stuff behaves differently on Kubernetes
(staging+production) than on Docker (dev+integration-test). I don't
see much value in fiddling with seccomp and capabilities to make
Docker mimic Kubernetes, because it's a lost cause. Smarter would
be to integration-test in a mini-Kubernetes cluster.

All that to say: let's try and deploy on staging, *then* learn how
our integration-test suite is flawed. To do that, we need tests
to pass :).