parse.csv: parse "" to NULL, not np.nan
In Pandas we use np.nan and not NULL. In Arrow, it's the opposite. We
have no meaning for np.nan, so we don't allow Arrow tables to use it.

Bonus feature: we now store int8 columns that include nulls as int8, not
float64. This is of dubious benefit today, since when we convert to
Pandas we'll switch back to float64. But it's nice to _store_ data in
the ideal format, at least. And in the future, if more modules eschew
Pandas, they'll be more efficient with nullable int8 columns.

Fixes error on production:

Module produced invalid data: arrow-validate:
--check-floats-all-finite failed on column A