Make lesson links relative ... and correct.
I spotted styling inconsistencies and incorrect links in English. I
guessed that the same occurred in Greek, and that the fixes are the
same.

Now, all inter-lesson links are relative (e.g.,
`<a href="scrape-using-xpath">`. This wasn't an option for us
previously, when lessons URLs used Django's oddball URL scheme (in which
each URL ends in `/`). But it's an option now, and I'm glad to declare
relative URLs the de-facto means of linking from one lesson to another.