Unit test python/django extraction; nix support for context in trans
Support for context in trans was interfering with message parameters.
Since we don't use it (and are not planning to use it),
I removed the support for it.

Now the behaviour of extraction scripts is unit tested.
Also, I18nMessage type in cjwkernel has a trans function
which is extracted normally.

trans_html was renamed to localize_html,
in order to emphasize its different nature from trans;
it's not expected to be parsed by pybabel
and it expects a locale argument.