Remove a signal receiver that creates problems
When allauth saved a new user during social login,
the deleted receiver would create a default user profile
(hence with the default locale id).
That default profile would be used later, instead of the updated one,
when trying to read the user profile and write the cookie locale,
resulting in the user seeing the default locale after login
and not the locale in their (updated) profile.