fetcher.fetch: fix unit tests
Also, handle migrate_params error gracefully and unit-test it.

We still have the problem that migrate_params() locks the workflow while
invoking the kernel.  We'll want to fix that. But this problem is in
`cjwstate.params.get_migrated_params()`, not `fetcher.fetch()`, and we
shouldn't try to solve it just in `fetcher.fetch()`.