Hard-code kubectl version in advanced-deploy
If we don't do this, kubectl 1.13 will be used by cloudbuild. It can't
set init-container images. I don't know why we didn't experience this
problem prior to the new cluster ... maybe because the cluster itself
was >1.13, and our new clusters are only 1.13? .... if this doesn't
work, we'll need to upgrade the clusters.