Split work into several node pools:
* main-pool-v2: standard 2-vCPU (sized to fit)
* web-pool: no-SMT (secure, and unfortunately 4-vCPU because 2 isn't enough)
* worker-pool: preemptible, big, cheap, no-SMT (secure)

The big expense: web servers run migrate_params(), which forces them to use
a sandbox. Sigh. If cjwkernel were an RPC service, perhaps, things might be
cheaper?

Anwyay, this should give savings because pre-emptible machines are ~25% the
price of non-pre-emptible. But we're at 32 vCPUs total now (and deployment
shouldn't stall); and we used to be at 24 vCPUs.