pandas.parse.csv: a little less Pandas...
The goal is zero Pandas in the parse codebase. This nixes Pandas from
the "detect whether a column is all '' or None" algorithm.

There's a change in behavior: an all-null column becomes text, not
number. This doesn't crop up often in the wild, I imagine; and when it
does, a number column seems a bizarre default.