pg_render_locker: avoid CancelledError during error
I searched and searched the Python docs, and I can't understand what
happens if a __aexit__() method raises asyncio.CancelledError. Maybe
nobody knows?

Anyway, in the context of PgRenderLocker, the only reasons to call
__aexit()__ are unit-test end and crash. This change doesn't make unit
tests crash, and it can only _help_ during a crash.

[refs #169843883] because there's a CancelledError somewhere in our logs
and I don't know what happens to it.