Bump parquet-to-arrow
[finishes #174867705]

This uses Arrow v1.0.1, but it emits 0.16.0-compatible files.

It also emits slightly-different metadata. In some tests,
cjwkernel.types.ArrowTable.__eq__() was giving false negatives in unit
tests. The solution: nix it entirely and use arrow_table_equals() (in
unit tests) and cjwparquet.are_files_equal() (when notifying).

We're well on our way to removing ArrowTable.table entirely -- and thus,
our dependency on pyarrow! (Outside of module code, that is.) The only
places we use pyarrow any more are:

* views.WfModule.value_counts()
* ArrowTable factories (reading metadata from file)
* cjwkernel.validate.*

It would be lovely to stop using pyarrow. A simpler web server is a
safer web server.