minio: use list_objects v1, not v2
This should make us compatible with GCS's S3 API. So we can configure
minio to use the S3 API, instead of the GCS API. [refs #176193903]

May solve [refs #175945954] and [#175975506]

We still use minio, even on production, for its IAM STS layer. One
alternative would be to rely on resumable file uploads and use a
presigned key. But that would need different client-side code on GCS
(production) than on minio/S3 (dev): something we'd like to avoid.