deploy/kubernetes/minio/init.sh: one role per service
[refs #176193903]. We eventually want to get here. We'll end up with a
bit of a bifurcated API: "minio" for minio-specific stuff (STS,
multipart transfers, multi-delete) and then ... uh ... "storage" for
generic operations (GET, PUT, DELETE). To transition, we'll need both
sets of credentials on many servers.

I think the biggest reason to change is speed. In particular,
multi-DELETE can be 30x faster if we allow 30 threads. And startup can
be 10x faster if we download all modules concurrently. [But we don't
need the speed quite yet, so I'm putting this all on hold.

For now, the GCS roles are set, and secrets are in Kubernetes. The
secrets aren't being used.