OutputIframe.js: nix `visible` prop
I think this was cruft from when we didn't use `key` to force a fresh
render of `DataGrid`. The idea was: when switching to a new Step, the
disappearance of the iframe could change the height of the DataGrid. We
no longer care, because if we switch to a new Step we render a new
DataGrid anyway.