Nix staticmodules [finishes #176717653]
Nix `parameters_version` from the module spec. It was a hack to give
internal modules "version numbers". Ten years from now we might want
something like it ... but no external modules use it and it isn't
documented, so we're best deleting it for now.

All modules are now i18n-ready. And the module-subsystem init is
simpler now.