Add "staticfilesdev" server in dev mode
This speeds up deploy: we "collectstatic" during build, so the actual
deploy phase is a mere "aws s3 cp".

And it makes dev mode more realistic: static files are all cross-origin
(localhost:8003), same as production.

The implementation is awkward because Django staticfiles makes the easy
stuff easier and the hard stuff harder. Ideally, we'll nix django-admin.