result-table-slice.json: just an Array, thanks
I don't understand how the JS side didn't crash before. The empty-table
shim was returning a JS Array, but the server was returning an Object
with a "rows" property. Also, the server would sometimes return 0 rows
when the client assumed 200, and that led to an infinite loop of
requests.

This change is simple. It's also sorta temporary, because "tile" is the
future, once we nix react-data-grid.