Change embed-data URL [refs #177350226]
The new URL's output can be cached in the browser, since it includes the
delta ID.

There's a new mechanism to pass data to the HTML: `?dataUrl=...`. The
HTML must respect this new URL; it shouldn't build a URL based on its
own URL. For compatibility, old-style modules will work: building a URL
based on our own URL is okay _for today_. Deploy plan is:

1. Deploy this new code
2. Update all modules, on production
3. Deploy new module-HTML URLs

After this deploy, the "hashchange" event won't fire. Iframes will
always be reloaded when data changes. There's a new mechanism, the
"subscribe-to-data-url" MessageEvent, that upgraded modules can use to
recover from the regression.