177350226 urls with workflow (#240)
* New embed URL [refs #177350226]

This URL also has a new feature: it queues a render if needed.
This was an easy win. [finishes #177420781]

* Nix "/404/" and "/403/" URLs

They seem to have only been used for simulating such errors: "404" on
dev (it's annoying to simulate because Django hides it in DEBUG mode)
and "403" on production (I can't think of a justification for this).

* Nix 'xyzzy/signup/' URL

It's been years. We don't need it.

* integrationtests: make pythoncode iframe code robust