Revamp uploaded files
`pandas_v0` no longer handles uploaded files. Only `arrow_v0` does.

We use a new Thrift datatype, `UploadedFile`, to hold the original
(user-supplied, unsafe) filename and the uploaded timestamp. The
`arrow_v0` framework converts to `Path`, dropping this data, for
backwards-compatibility with existing modules.

The win: now, the params passed to a module are valid -- and they are
the same as the params selected on the client side.