176951900 date column (#246)
Revamp Thrift protocol and kernel

* Move `ModuleSpec`, `ParamField` and `ParamSchema` to `cjwmodule` package. Now, modules' unit tests can use a pattern that prevents invalid params.
* Make the `params` dictionary we pass to a module exactly the same as the `params` seen in the UI (when valid). Previously, there were subtle differences that made for inconsistent module code (Pandas "tab" param was a complex object; Arrow "tab" param was an str).
* Make `cjwmodule.pandas.framework` package, with `pandas_v0`, `arrow_v0` and (upcoming) `arrow_v1` frameworks. The framework handles conversion to/from Thrift messages.
* Nix Thrift `TableMetadata`: it was usually redundant (a "pyarrow.string" column has Workbench type "text", always) and this new convention gives a way for Arrow modules to output number formats and date resolutions.

Finally, we can write Arrow modules that output dates.

Squashed merge of the following:

* cjwkernel: support Date columns [refs #176951900]

* Upgrade to Django 3.2 (freezing auto-upgraded us)
* cjwpandasmodule: put some pandas<=>arrow conversions here
* cjwkernel.types.ColumnType: convert to NamedTuple
* nix get_stored_dataframe, unused by all modules

* Support date columns [refs #176951900]

This passes tests, but there are no modules that actually _use_ date
columns and so I haven't actually tested it.

* cjwmodule~=4.0

Moves some data structures out of `cjwkernel.types`.

Coming up:

* A module's `render_arrow_v1()` will introduce Yet Another Format for
  modules. Our first standardized API. (Its return value will be a
  `cjwmodule.arrow.types.ArrowRenderResult`: no coercing, just
  validating.)
* Some aspect of module-spec validation and params will move to
  cjwmodule. This will let modules' unit-test suites ensure module
  authors write valid params.
* ... and then we'll add a whole bunch of modules, to support dates!

* docker-compose.yml: Fix tusd config in dev mode

* Fix unit tests

* Nearly nix ArrowTable

This is a new Thrift message-passing convention: instead of passing a
schema separately from the Arrow table, we pass the schema inside the
Arrow table. (Some fields have "format" or "unit" metadata; that's it.)

Now it's possible for an Arrow module to create a formatted number
column ... or a date column [#176951900].

* Move ArrowTable to cjwkernel.pandas.module

* cjwkernel: introduce the idea of "framework"

Now, we can convert thrift<=>pandas directly instead of going
thrift<=>arrow<=>pandas. The double-conversion idea duplicated all our
bad decisions. For example: Arrow modules got the
"empty-fetched-file-is-empty-parquet-dataframe" feature ... even though
we don't want that feature.)

With each commit, I'm reducing the number of conversion functions and
hops. The big prize is Params....

* arrow_v0: fix infer format from new column

* cjwstate.modules => cjwmodule.spec

Unit tests pass, but I haven't tested more extensively than that.

* Typo fix caught by integration tests

* thrift.types: nix "Params". Use Python dict instead

The distinction between "Params" and "dict" was too subtle. It ended up
being a lot of code just to turn a "Tab" parameter into a "TabOutput"
... and with the advent of module frameworks, we have a better way to do
that.

* Integration tests: resolve pythoncode race (I think)

* frameworks: fix recursion error with ParamSchema.Map

* Bump cjwmodule

* Revamp uploaded files

`pandas_v0` no longer handles uploaded files. Only `arrow_v0` does.

We use a new Thrift datatype, `UploadedFile`, to hold the original
(user-supplied, unsafe) filename and the uploaded timestamp. The
`arrow_v0` framework converts to `Path`, dropping this data, for
backwards-compatibility with existing modules.

The win: now, the params passed to a module are valid -- and they are
the same as the params selected on the client side.

* integrationtests.test_pythoncode: typo fix

* Another integrationtest-inspired fix

* Organize preloads

* One last cjwmodule update, for timezone param fix