kernel: Spin 10s, not 1s, wait()-ing on a child
[refs #178218352] We were getting timeouts from the Check module. But in
on-production testing, computations only last a fraction of the timeout.

So I suspect the timeout happened when reaping. Since the problem only
happens on production, let's debug it on production! I'll make the
timeout 10s longer. If the timeouts disappear, then evidence suggests
our 1s limit was too short.

The timeout is time between pipe closed and process exited. In
retrospect, 1s seems a bit optimistic.