Bump max number columns to 600
We were about to hit a limit on
"CSSEGISandData / COVID-19 time-series global", which
tracks COVID with one column (instead of row) per date.

What a frustrating data source! It is both very important and very
inefficient. We may want to add a "Transpose" checkbox to our CSV
reader; but then we'd still be stuck with 300 columns, one per
country+province.